// de-care/libs/domains/account/state-account/src/lib/data-services/account.enums.ts
export enum CustomerTypeEnum {
    NewAccount = 'NEW_ACCOUNT',
    NewSxirAccount = 'NEW_SXIR_ACCOUNT',
    InTrial = 'IN_TRIAL',
    InTrialStreaming = 'SXIR_IN_TRIAL',
    PostTrial = 'POST_TRIAL',
    WBGA = 'WBGA',
    WBSP = 'WBSP',
    TrialActivation = 'TRIAL_ACTIVATION',
    NewSxirStudent = 'SXIR_NEW_STUDENT',
    SelfPay = 'SELF_PAY',
    SelfPayStreaming = 'SXIR_SELFPAY',
    ReverifyStudent = 'SXIR_REVERIFY_STUDENT',
    SxirRtdTrial = 'SXIR_RTD',
    RtpTrialActivation = 'RTP_TRIAL_ACTIVATION',
    SxirAddSubscription = 'SXIR_ADD_SUBSCRIPTION',
}

export enum PlanTypeEnum {
    Trial = 'TRIAL',
    Introductory = 'INTRODUCTORY',
    Promo = 'PROMO',
    Courtesy = 'COURTESY',
    Demo = 'DEMO',
    Unset = 'UNSET',
    SelfPaid = 'SELF_PAID',
    SelfPay = 'SELF_PAY',
    PromoMCP = 'PROMO_MCP',
    TrialExtension = 'TRIAL_EXT',
    TrialExtensionRTC = 'TRIAL_EXT_RTC',
    RtpOffer = 'RTP_OFFER',
    TrialRtp = 'TRIAL_RTP',
    TrialGgle = 'TRIAL_GGLE',
    LongTerm = 'LONG_TERM',
    Advantage = 'ADVANTAGE',
    Default = 'DEFAULT',
    RtdTrial = 'RTD_OFFER',
}

export enum BillingStatusEnum {
    CC_Current = 'CC_CURRENT',
    CC_Next = 'CC_NEXT',
    CC_Current_Next = 'CC_CURRENT_NEXT',
    CC_Suspended = 'CC_SUSPENDED',
    INV_Current = 'INV_CURRENT',
    INV_Next = 'INV_NEXT',
    INV_Current_Next = 'INV_CURRENT_NEXT',
    INV_Suspended = 'INV_SUSPENDED',
}